refactor: rename Doc to Sampler and move it to lib
Rename the Doc class to Sampler.

BREAKING CHANGE: the exported class is named Sampler and not Doc anymore
